Update KotlinTest version
This commit updates KotlinTest to it's latest version (current version is 2.0.1, which is a release from 3 years ago).

To do that this commit updates the `jars` folder to all current KotlinTest dependencies, downloaded from MavenCentral.

KotlinTest doesn't use JUnit, so `cyber-dojo.sh` was updated to use KotlinTest's runner. With a caveat: The test class name must remain as `HikerTest`, and the package of that class must remain `hiker`. This was commented in the source file itself so that users don't break the execution.

This is necessary because a KotlinTest project usually executes with Gradle or Maven, and these tools know how to find all the classes and execute them one by one. As we're using command-line only, we don't have that kind of intelligence out-of-the-box, so I decided to hard-code it.

Due to the change above, the `red_amber_green.rb` file had to be changed, as KotlinTest's regex is different from JUnit's.

Finally, the `HikerTest` was updated to reflect a more realistic KotlinTest execution with `StringSpec`.