Include hard-wired-test-name file at visible_filenames
There are some warnings at the files "`[X]`" which indicate looking for a text file with a description. However, this text file is invisible at the workspace.

I believe this is due to `manifest.json` missing and indication for it, and for this reason I added it to the file.